requirements.txt
asgiref==3.5.2
Django==4.1.4
gunicorn==20.1.0
psycopg2==2.9.5
sqlparse==0.4.3
djangorestframework-simplejwt
djangorestframework~=3.14.0
pydantic~=1.10.2